# admin-api/swagger/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Admin panel API",
    "description": "It's API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api"
    }
  ],
...
  "paths": {
...
    "/content/playlist/{id}": {
      "get": {
        "tags": [
          "playlist"
        ],
        "summary": "Get playlist by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistInfo"
                }
              }
            }
          },
          "400": {
            "description": "invalid query",
            "content": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "playlist"
        ],
        "summary": "Deletes a playlist by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistInfo"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query",
            "content": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PlaylistInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "user_id": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string"
          },
          "img_src": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}